Win32 makefile updates and minor fix to half floats for double/float conversions.
git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@2448 afdca40c-03d6-0310-8ede-e9f093b21075

+++ src/ShHalfImpl.hpp
@@ -178,7 +178,7 @@ ShHalf::T ShHalf::to_val(double value) {
   // INF in value, and is rather slow. 
   result = (sign << 15);
   if(fraction == 0) { // zero
-  } else if(fabsf(value) > 65504) { // INF
+  } else if(fabs(value) > 65504) { // INF
     result |= 31 << 10;
   } else if(exponent < -13) { // denormalized 
     int significand = (int)(fraction * (1LL << (exponent + 24)));
